Renamed directories and added new repositories
