+++ src/main/java/org/example/alphasolutions/controllers/ManagementSoftwareController.java
@@ -1,8 +1,8 @@
-package org.example.alphasolutions.Controller;
+package org.example.alphasolutions.controllers;
 
 import jakarta.servlet.http.HttpSession;
-import org.example.alphasolutions.Model.*;
-import org.example.alphasolutions.Service.ManagementSoftwareService;
+import org.example.alphasolutions.models.*;
+import org.example.alphasolutions.services.ManagementSoftwareService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
