+++ src/main/java/org/example/alphasolutions/models/Priority.java
@@ -1,4 +1,4 @@
-package org.example.alphasolutions.Model;
+package org.example.alphasolutions.models;
 
 public enum Priority {
     HIGH,
