+++ src/main/java/org/example/alphasolutions/models/Status.java
@@ -1,4 +1,4 @@
-package org.example.alphasolutions.Model;
+package org.example.alphasolutions.models;
 
 public enum Status {
     COMPLETED,
