+++ src/main/java/org/example/alphasolutions/models/Task.java
@@ -1,4 +1,4 @@
-package org.example.alphasolutions.Model;
+package org.example.alphasolutions.models;
 
 import java.util.Date;
